refactor models dir (#337)
